refactor(deeptrace): modularize memory management
