--- crates/ethernity-deeptrace/src/memory.rs
@@ -1,333 +1,0 @@
-use parking_lot::RwLock;
-use std::collections::HashMap;
-use std::num::NonZeroUsize;
-use std::sync::Arc;
-use std::time::{Duration, Instant};
-
-/// Módulo para gerenciamento de memória e performance
-pub mod memory {
-    use super::*;
-
-    /// Cache LRU para dados frequentemente acessados
-    pub struct SmartCache<K, V>
-    where
-        K: std::hash::Hash + Eq + Clone,
-        V: Clone,
-    {
-        cache: RwLock<lru::LruCache<K, CacheEntry<V>>>,
-        stats: RwLock<CacheStats>,
-        ttl: Duration,
-    }
-
-    /// Entrada de cache com timestamp
-    struct CacheEntry<V> {
-        value: V,
-        expires_at: Instant,
-    }
-
-    /// Estatísticas de cache
-    #[derive(Debug, Default, Clone)]
-    pub struct CacheStats {
-        pub hits: usize,
-        pub misses: usize,
-        pub inserts: usize,
-        pub evictions: usize,
-        pub expirations: usize,
-    }
-
-    impl<K, V> SmartCache<K, V>
-    where
-        K: std::hash::Hash + Eq + Clone,
-        V: Clone,
-    {
-        /// Cria um novo cache com capacidade e TTL especificados
-        pub fn new(capacity: usize, ttl: Duration) -> Self {
-            let capacity = NonZeroUsize::new(capacity).unwrap_or(NonZeroUsize::new(1).unwrap());
-
-            Self {
-                cache: RwLock::new(lru::LruCache::new(capacity)),
-                stats: RwLock::new(CacheStats::default()),
-                ttl,
-            }
-        }
-
-        /// Obtém um valor do cache
-        pub fn get(&self, key: &K) -> Option<V> {
-            let mut cache = self.cache.write();
-            let now = Instant::now();
-
-            if let Some(entry) = cache.get(key) {
-                if entry.expires_at > now {
-                    // Cache hit
-                    self.stats.write().hits += 1;
-                    Some(entry.value.clone())
-                } else {
-                    // Entry expired
-                    cache.pop(key);
-                    self.stats.write().expirations += 1;
-                    None
-                }
-            } else {
-                // Cache miss
-                self.stats.write().misses += 1;
-                None
-            }
-        }
-
-        /// Insere um valor no cache
-        pub fn insert(&self, key: K, value: V) {
-            let mut cache = self.cache.write();
-            let expires_at = Instant::now() + self.ttl;
-
-            if cache.put(key, CacheEntry { value, expires_at }).is_some() {
-                self.stats.write().evictions += 1;
-            }
-
-            self.stats.write().inserts += 1;
-        }
-
-        /// Obtém estatísticas do cache
-        pub fn stats(&self) -> CacheStats {
-            self.stats.read().clone()
-        }
-    }
-
-    /// Pool de buffers para reutilização
-    pub struct BufferPool {
-        buffers: RwLock<Vec<Vec<u8>>>,
-        buffer_size: usize,
-        max_buffers: usize,
-        stats: RwLock<BufferPoolStats>,
-    }
-
-    /// Estatísticas do pool de buffers
-    #[derive(Debug, Default, Clone)]
-    pub struct BufferPoolStats {
-        pub allocations: usize,
-        pub reuses: usize,
-        pub returns: usize,
-        pub misses: usize,
-    }
-
-    impl BufferPool {
-        /// Cria um novo pool de buffers
-        pub fn new(buffer_size: usize, max_buffers: usize) -> Self {
-            Self {
-                buffers: RwLock::new(Vec::with_capacity(max_buffers)),
-                buffer_size,
-                max_buffers,
-                stats: RwLock::new(BufferPoolStats::default()),
-            }
-        }
-
-        /// Obtém um buffer do pool ou cria um novo
-        pub fn get_buffer(&self) -> Vec<u8> {
-            let mut buffers = self.buffers.write();
-
-            if let Some(mut buffer) = buffers.pop() {
-                // Limpa o buffer antes de reutilizar
-                buffer.clear();
-                self.stats.write().reuses += 1;
-                buffer
-            } else {
-                // Cria um novo buffer
-                self.stats.write().allocations += 1;
-                self.stats.write().misses += 1;
-                Vec::with_capacity(self.buffer_size)
-            }
-        }
-
-        /// Devolve um buffer ao pool
-        pub fn return_buffer(&self, mut buffer: Vec<u8>) {
-            let mut buffers = self.buffers.write();
-
-            // Só adiciona ao pool se houver espaço
-            if buffers.len() < self.max_buffers {
-                // Limpa o buffer antes de devolver
-                buffer.clear();
-                buffers.push(buffer);
-                self.stats.write().returns += 1;
-            }
-        }
-
-        /// Obtém estatísticas do pool
-        pub fn stats(&self) -> BufferPoolStats {
-            self.stats.read().clone()
-        }
-    }
-
-    /// Gerenciador de memória para a workspace
-    pub struct MemoryManager {
-        caches: RwLock<HashMap<String, Arc<dyn std::any::Any + Send + Sync>>>,
-        buffer_pools: RwLock<HashMap<String, Arc<BufferPool>>>,
-    }
-
-    impl MemoryManager {
-        /// Cria um novo gerenciador de memória
-        pub fn new() -> Self {
-            Self {
-                caches: RwLock::new(HashMap::new()),
-                buffer_pools: RwLock::new(HashMap::new()),
-            }
-        }
-
-        /// Registra um cache
-        pub fn register_cache<K, V>(&self, name: &str, cache: Arc<SmartCache<K, V>>)
-        where
-            K: std::hash::Hash + Eq + Clone + Send + Sync + 'static,
-            V: Clone + Send + Sync + 'static,
-        {
-            self.caches.write().insert(name.to_string(), cache as Arc<dyn std::any::Any + Send + Sync>);
-        }
-
-        /// Registra um pool de buffers
-        pub fn register_buffer_pool(&self, name: &str, pool: Arc<BufferPool>) {
-            self.buffer_pools.write().insert(name.to_string(), pool);
-        }
-
-        /// Obtém estatísticas de uso de memória
-        pub fn memory_usage(&self) -> MemoryUsageStats {
-            let mut stats = MemoryUsageStats::default();
-
-            // Coleta estatísticas de caches
-            for (name, _cache) in self.caches.read().iter() {
-                // Simplificado - apenas conta o número de caches
-                stats.cache_stats.insert(name.clone(), CacheStatsInfo {
-                    hits: 0,
-                    misses: 0,
-                    hit_ratio: 0.0,
-                    entries: 0,
-                });
-            }
-
-            // Coleta estatísticas de pools de buffers
-            for (name, pool) in self.buffer_pools.read().iter() {
-                let pool_stats = pool.stats();
-                stats.buffer_pool_stats.insert(name.clone(), BufferPoolStatsInfo {
-                    allocations: pool_stats.allocations,
-                    reuses: pool_stats.reuses,
-                    reuse_ratio: if pool_stats.allocations > 0 {
-                        pool_stats.reuses as f64 / (pool_stats.allocations + pool_stats.reuses) as f64
-                    } else {
-                        0.0
-                    },
-                });
-            }
-
-            stats
-        }
-    }
-
-    /// Estatísticas de uso de memória
-    #[derive(Debug, Default, Clone)]
-    pub struct MemoryUsageStats {
-        pub cache_stats: HashMap<String, CacheStatsInfo>,
-        pub buffer_pool_stats: HashMap<String, BufferPoolStatsInfo>,
-    }
-
-    /// Informações de estatísticas de cache
-    #[derive(Debug, Clone)]
-    pub struct CacheStatsInfo {
-        pub hits: usize,
-        pub misses: usize,
-        pub hit_ratio: f64,
-        pub entries: usize,
-    }
-
-    /// Informações de estatísticas de pool de buffers
-    #[derive(Debug, Clone)]
-    pub struct BufferPoolStatsInfo {
-        pub allocations: usize,
-        pub reuses: usize,
-        pub reuse_ratio: f64,
-    }
-
-    /// Monitorador de uso de memória
-    pub struct MemoryMonitor {
-        memory_manager: Arc<MemoryManager>,
-        sampling_interval: Duration,
-        history: Arc<RwLock<Vec<MemoryUsageSnapshot>>>, // Now wrapped in Arc
-        max_history: usize,
-    }
-
-    /// Snapshot de uso de memória
-    #[derive(Debug, Clone)]
-    pub struct MemoryUsageSnapshot {
-        pub timestamp: std::time::SystemTime,
-        pub stats: MemoryUsageStats,
-        pub system_memory: SystemMemoryInfo,
-    }
-
-    /// Informações de memória do sistema
-    #[derive(Debug, Clone, Default)]
-    pub struct SystemMemoryInfo {
-        pub total_memory: u64,
-        pub used_memory: u64,
-        pub free_memory: u64,
-    }
-
-    impl MemoryMonitor {
-        /// Cria um novo monitorador de memória
-        pub fn new(memory_manager: Arc<MemoryManager>, sampling_interval: Duration, max_history: usize) -> Self {
-            Self {
-                memory_manager,
-                sampling_interval,
-                history: Arc::new(RwLock::new(Vec::with_capacity(max_history))),
-                max_history,
-            }
-        }
-
-        /// Inicia o monitoramento
-        pub async fn start_monitoring(&self) -> Result<(), ()> {
-            let memory_manager = self.memory_manager.clone();
-            let sampling_interval = self.sampling_interval;
-            let history = self.history.clone();
-            let max_history = self.max_history;
-
-            tokio::spawn(async move {
-                loop {
-                    // Coleta estatísticas
-                    let stats = memory_manager.memory_usage();
-                    let system_memory = get_system_memory_info();
-
-                    let snapshot = MemoryUsageSnapshot {
-                        timestamp: std::time::SystemTime::now(),
-                        stats,
-                        system_memory,
-                    };
-
-                    // Adiciona ao histórico
-                    {
-                        let mut history_guard = history.write();
-                        history_guard.push(snapshot);
-
-                        // Limita o tamanho do histórico
-                        if history_guard.len() > max_history {
-                            history_guard.remove(0);
-                        }
-                    }
-                    // Espera até o próximo intervalo
-                    tokio::time::sleep(sampling_interval).await;
-                }
-            });
-
-            Ok(())
-        }
-
-        /// Obtém o histórico de uso de memória
-        pub fn get_history(&self) -> Vec<MemoryUsageSnapshot> {
-            self.history.read().clone() // Clone the inner Vec
-        }
-    }
-
-    /// Obtém informações de memória do sistema
-    fn get_system_memory_info() -> SystemMemoryInfo {
-        // Implementação simplificada - retorna valores padrão
-        // Em um ambiente real, você usaria sysinfo
-        SystemMemoryInfo {
-            total_memory: 0,
-            used_memory: 0,
-            free_memory: 0,
-        }
-    }
-}
