+++ crates/ethernity-deeptrace/src/memory/monitor.rs
@@ -1,0 +1,93 @@
+use super::{MemoryManager, MemoryUsageStats};
+use parking_lot::RwLock;
+use std::sync::Arc;
+use std::time::Duration;
+
+/// Monitorador de uso de memória
+pub struct MemoryMonitor {
+    memory_manager: Arc<MemoryManager>,
+    sampling_interval: Duration,
+    history: Arc<RwLock<Vec<MemoryUsageSnapshot>>>,
+    max_history: usize,
+}
+
+/// Snapshot de uso de memória
+#[derive(Debug, Clone)]
+pub struct MemoryUsageSnapshot {
+    pub timestamp: std::time::SystemTime,
+    pub stats: MemoryUsageStats,
+    pub system_memory: SystemMemoryInfo,
+}
+
+/// Informações de memória do sistema
+#[derive(Debug, Clone, Default)]
+pub struct SystemMemoryInfo {
+    pub total_memory: u64,
+    pub used_memory: u64,
+    pub free_memory: u64,
+}
+
+impl MemoryMonitor {
+    /// Cria um novo monitorador de memória
+    pub fn new(memory_manager: Arc<MemoryManager>, sampling_interval: Duration, max_history: usize) -> Self {
+        Self {
+            memory_manager,
+            sampling_interval,
+            history: Arc::new(RwLock::new(Vec::with_capacity(max_history))),
+            max_history,
+        }
+    }
+
+    /// Inicia o monitoramento
+    pub async fn start_monitoring(&self) -> Result<(), ()> {
+        let memory_manager = self.memory_manager.clone();
+        let sampling_interval = self.sampling_interval;
+        let history = self.history.clone();
+        let max_history = self.max_history;
+
+        tokio::spawn(async move {
+            loop {
+                // Coleta estatísticas
+                let stats = memory_manager.memory_usage();
+                let system_memory = get_system_memory_info();
+
+                let snapshot = MemoryUsageSnapshot {
+                    timestamp: std::time::SystemTime::now(),
+                    stats,
+                    system_memory,
+                };
+
+                // Adiciona ao histórico
+                {
+                    let mut history_guard = history.write();
+                    history_guard.push(snapshot);
+
+                    // Limita o tamanho do histórico
+                    if history_guard.len() > max_history {
+                        history_guard.remove(0);
+                    }
+                }
+                // Espera até o próximo intervalo
+                tokio::time::sleep(sampling_interval).await;
+            }
+        });
+
+        Ok(())
+    }
+
+    /// Obtém o histórico de uso de memória
+    pub fn get_history(&self) -> Vec<MemoryUsageSnapshot> {
+        self.history.read().clone()
+    }
+}
+
+/// Obtém informações de memória do sistema
+fn get_system_memory_info() -> SystemMemoryInfo {
+    // Implementação simplificada - retorna valores padrão
+    SystemMemoryInfo {
+        total_memory: 0,
+        used_memory: 0,
+        free_memory: 0,
+    }
+}
+
